Split green into submodules
