+++ src/green/element.rs
@@ -1,0 +1,37 @@
+use super::*;
+use crate::{cursor::SyntaxKind, NodeOrToken, TextUnit};
+
+pub type GreenElement = NodeOrToken<GreenNode, GreenToken>;
+
+impl From<GreenNode> for GreenElement {
+    #[inline]
+    fn from(node: GreenNode) -> GreenElement {
+        NodeOrToken::Node(node)
+    }
+}
+
+impl From<GreenToken> for GreenElement {
+    #[inline]
+    fn from(token: GreenToken) -> GreenElement {
+        NodeOrToken::Token(token)
+    }
+}
+
+impl GreenElement {
+    /// Returns kind of this element.
+    #[inline]
+    pub fn kind(&self) -> SyntaxKind {
+        match self {
+            NodeOrToken::Node(it) => it.kind(),
+            NodeOrToken::Token(it) => it.kind(),
+        }
+    }
+    /// Returns length of the text covered by this element.
+    #[inline]
+    pub fn text_len(&self) -> TextUnit {
+        match self {
+            NodeOrToken::Node(it) => it.text_len(),
+            NodeOrToken::Token(it) => it.text_len(),
+        }
+    }
+}
