+++ src/lib.rs
@@ -9,6 +9,7 @@
 #![deny(unsafe_code)]
 #![allow(unused)]
 
+#[warn(unused)]
 mod green;
 #[allow(unsafe_code)]
 pub mod cursor;
